fastapi_graphql_client/requirements.txt
fastapi
uvicorn
strawberry-graphql
grpcio
grpcio-tools